+++ director/interfaces.go
@@ -1,5 +1,6 @@
 package director
 
+// Director - interface type of game director
 type Director interface {
 	PlayScene(command SceneRequest) Result
 }