+++ scg/expr/errors.go
@@ -5,9 +5,10 @@ import (
 )
 
 var (
-	StartSceneNotFoundError = errors.New("start scene with it name not found")
+	StartSceneNotFoundError   = errors.New("start scene with it name not found")
+	GoodbyeSceneNotFoundError = errors.New("goodbye scene with it name not found")
 )
 
-func errorNameSceneExists(sceneName string) error {
-	return errors.New("this name scene already exist \"" + sceneName + "\"")
+func errorNameSceneNotFound(sceneName string) error {
+	return errors.New("this name scene not found: \"" + sceneName + "\"")
 }