+++ scg/expr/expr.go
@@ -17,13 +17,22 @@ type ScriptInfo struct {
 }
 
 func (si *ScriptInfo) IsValid() (is bool, err error) {
+	unknownScene := ""
 	for _, sc := range si.Script {
 		if is, err = sc.IsValid(); !is {
 			break
 		}
+		for _, name := range sc.NextScenes {
+			if _, is = si.Script[name]; !is {
+				unknownScene = name
+			}
+		}
 	}
 	if !is {
-		return is, err
+		if unknownScene == "" {
+			return is, err
+		}
+		return is, errorNameSceneNotFound(unknownScene)
 	}
 
 	if _, is = si.Script[si.GoodByeScene]; !is {