+++ scg/script/codegen/script.go
@@ -13,9 +13,16 @@ import (
 // ScriptFile returns structs for script
 func ScriptFile(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) []*codegen.File {
 	directorConfigFile := directorConfig(rootPkg, rootDir, scriptInfo)
-	scriptFile := scriptScenes(rootPkg, rootDir, scriptInfo)
+	scriptFiles := make([]*codegen.File, 0)
+	for key, value := range scriptInfo.Script {
+		scriptFiles = append(scriptFiles, scriptScenes(rootPkg, rootDir, scriptInfo.Name, sceneWithName{
+			Scene: value,
+			Name:  key,
+		}))
+	}
+	sceneNamesFile := sceneNames(rootPkg, rootDir, scriptInfo)
 
-	return []*codegen.File{scriptFile, directorConfigFile}
+	return append([]*codegen.File{directorConfigFile, sceneNamesFile}, scriptFiles...)
 }
 
 func directorConfig(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) *codegen.File {
@@ -24,6 +31,7 @@ func directorConfig(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) 
 	fpath := filepath.Join(rootDir, "script", "init.go")
 	imports := []*codegen.ImportSpec{
 		{Path: path.Join(rootPkg, "manager")},
+		{Path: path.Join(rootPkg, "script", "scenes")},
 		codegen.SCGNamedImport("director", "game"),
 	}
 
@@ -33,7 +41,7 @@ func directorConfig(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) 
 
 	sections = append(sections, &codegen.SectionTemplate{
 		Name:   "director-config",
-		Source: directorConfigStruct,
+		Source: directorConfigStructT,
 		Data:   scriptInfo,
 		FuncMap: map[string]interface{}{
 			"ToTitle": codegen.ToTitle,
@@ -41,7 +49,30 @@ func directorConfig(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) 
 		},
 	})
 
-	return &codegen.File{Path: fpath, SectionTemplates: sections}
+	return &codegen.File{Path: fpath, SectionTemplates: sections, IsUpdatable: true}
+}
+
+func sceneNames(_ string, rootDir string, scriptInfo expr.ScriptInfo) *codegen.File {
+	var sections []*codegen.SectionTemplate
+
+	fpath := filepath.Join(rootDir, "script", "scenes", "names.go")
+	var imports []*codegen.ImportSpec
+
+	sections = []*codegen.SectionTemplate{
+		codegen.Header(codegen.ToTitle(scriptInfo.Name)+"-Scenes Name", "scenes", imports, false),
+	}
+
+	sections = append(sections, &codegen.SectionTemplate{
+		Name:   "scenes-name",
+		Source: scenesConstantConfigStructT,
+		Data:   scriptInfo,
+		FuncMap: map[string]interface{}{
+			"ToTitle": codegen.ToTitle,
+			"ToSnake": codegen.SnakeCase,
+		},
+	})
+
+	return &codegen.File{Path: fpath, SectionTemplates: sections, IsUpdatable: true}
 }
 
 type sceneWithName struct {
@@ -49,41 +80,36 @@ type sceneWithName struct {
 	Name string
 }
 
-func scriptScenes(rootPkg string, rootDir string, scriptInfo expr.ScriptInfo) *codegen.File {
+func scriptScenes(rootPkg string, rootDir string, scriptName string, sceneInfo sceneWithName) *codegen.File {
 	var sections []*codegen.SectionTemplate
 
-	fpath := filepath.Join(rootDir, "script", "script.go")
+	fpath := filepath.Join(rootDir, "script", "scenes", codegen.SnakeCase(sceneInfo.Name)+"_scene.go")
 	imports := []*codegen.ImportSpec{
 		codegen.SCGImport(path.Join("director", "scene")),
 		codegen.SCGNamedImport(path.Join("director", "matchers"), "base_matchers"),
 		{Path: path.Join(rootPkg, "script", "matchers")},
 		{Path: path.Join(rootPkg, "script", "errors")},
+		{Path: path.Join(rootPkg, "script", "payloads")},
 		{Path: path.Join(rootPkg, "manager")},
 	}
 
 	sections = []*codegen.SectionTemplate{
-		codegen.Header(codegen.ToTitle(scriptInfo.Name)+"-SceneStructs", "script", imports, true),
+		codegen.Header(codegen.ToTitle(scriptName)+"-SceneStructs", "scenes", imports, true),
 	}
 
-	for key, value := range scriptInfo.Script {
-		sc := sceneWithName{
-			Scene: value,
-			Name:  key,
-		}
-		sections = append(sections, &codegen.SectionTemplate{
-			Name:   "scene-struct-" + key,
-			Source: sceneStructT,
-			Data:   sc,
-			FuncMap: map[string]interface{}{
-				"ToTitle":              codegen.ToTitle,
-				"CamelCase":            codegen.CamelCase,
-				"ConvertNameToMatcher": matchers.ConvertNameToMatcher,
-				"ConvertNameToError":   errors2.ConvertNameToError,
-			},
-		})
-	}
+	sections = append(sections, &codegen.SectionTemplate{
+		Name:   "scene-struct-" + sceneInfo.Name,
+		Source: sceneStructT,
+		Data:   sceneInfo,
+		FuncMap: map[string]interface{}{
+			"ToTitle":              codegen.ToTitle,
+			"CamelCase":            codegen.CamelCase,
+			"ConvertNameToMatcher": matchers.ConvertNameToMatcher,
+			"ConvertNameToError":   errors2.ConvertNameToError,
+		},
+	})
 
-	return &codegen.File{Path: fpath, SectionTemplates: sections}
+	return &codegen.File{Path: fpath, SectionTemplates: sections, IsUpdatable: true}
 }
 
 var ln = 0
@@ -98,38 +124,53 @@ func storeLen(l int) bool {
 	return false
 }
 
-const sceneStructT = `type {{ ToTitle .Name }} struct {
-		TextManager    manager.TextManager
-	}
-	
-	func (sc *{{ ToTitle .Name }}) React(_ *scene.Context) scene.Command {
-		// TODO
-		return scene.NoCommand
-	}
-	
-	func (sc *{{ ToTitle .Name }}) Next() scene.Scene {
-		//TODO
-		return &{{ ToTitle .Name }}{TextManager: sc.TextManager}
-	}
-	
-	func (sc *{{ ToTitle .Name }}) GetSceneInfo(ctx *scene.Context) (scene.Info, bool) {
-		{{ if .Text.Values }}var (
-			{{range $nameVar, $typeVar := .Text.Values}}{{$nameVar}} {{$typeVar}}
-			{{end}}
-		)
-		{{end}}
-		//TODO
+const sceneStructT = `// {{ ToTitle .Name }} scene
+type {{ ToTitle .Name }} struct {
+	TextManager    manager.TextManager
+	NextScene 	   SceneName
+}
 
-		text, _ := sc.TextManager.Get{{ ToTitle .Name }}Text(
-			{{range $nameVar, $typeVar := .Text.Values}}{{$nameVar}},
-			{{end}})
-		return scene.Info{
-			Text: text,
-			ExpectedMessages: []scene.MessageMatcher{ ` + sceneMatchersStructT + ` },
-			Buttons: []scene.Button{},
-			{{ if .Scene.Error.IsValid }}Err: ` + sceneErrorsStructT + `,{{end}}
-		}, true
+// React function of actions after scene has been played
+func (sc *{{ ToTitle .Name }}) React(_ *scene.Context) scene.Command {
+	// TODO Write the actions after {{ ToTitle .Name }} scene has been played
+
+	sc.NextScene = {{ ToTitle .Name }}Scene // TODO: manually set next scene after reaction
+	return scene.NoCommand
+}
+
+// Next function returning next scene
+func (sc *{{ ToTitle .Name }}) Next() scene.Scene {
+	{{ if .NextScenes }}switch sc.NextScene { 
+		{{ range .NextScenes }} case {{ ToTitle . }}Scene:
+			return &{{ ToTitle . }}{
+				TextManager: sc.TextManager,
+			}
+		{{end}}}{{end}}
+
+	return &{{ ToTitle .Name }}{
+			TextManager: sc.TextManager,
 	}
+}
+
+// Next function returning info about scene
+func (sc *{{ ToTitle .Name }}) GetSceneInfo(_ *scene.Context) (scene.Info, bool) {
+	{{ if .Text.Values }}var (
+		{{range $nameVar, $typeVar := .Text.Values}}{{$nameVar}} {{$typeVar}}
+		{{end}}
+	)
+	{{end}}
+	// TODO Write some actions for get data for texts
+
+	text, _ := sc.TextManager.Get{{ ToTitle .Name }}Text(
+		{{range $nameVar, $typeVar := .Text.Values}}{{$nameVar}},
+		{{end}})
+	return scene.Info{
+		Text: text,
+		ExpectedMessages: []scene.MessageMatcher{ ` + sceneMatchersStructT + ` },
+		Buttons: []scene.Button{ ` + sceneButtonsStructT + ` },
+		{{ if .Scene.Error.IsValid }}Err: ` + sceneErrorsStructT + `,{{end}}
+	}, true
+}
 `
 
 const sceneMatchersStructT = `{{ range .Matchers }}
@@ -138,18 +179,41 @@ const sceneMatchersStructT = `{{ range .Matchers }}
 	if .IsSelectMatcher }} matchers.{{ToTitle (.MustSelectsMatcher).Name}}Matcher,{{end}}{{end}} {{if .Matchers}}
 {{end}}`
 
+const sceneButtonsStructT = `{{ $sceneName := .Name }}{{ range $name, $button := .Buttons }}
+{ 
+	Title: "{{ $button.Text }}", {{ if $button.URL }}
+	URL: "{{ $button.URL }}", {{end}} {{ if $button.Payload }}
+	Payload: &payloads.{{ ToTitle $sceneName}}{{ ToTitle $name }}Payload{},{{end}}
+}, {{end}} {{ if .Buttons }}
+{{end}}`
+
 const sceneErrorsStructT = `{{with .Scene.Error}}{{ if .IsBase }} base_matchers.{{ConvertNameToError .Base}}{{end}}{{
 	if .IsText }} errors.{{ToTitle .Name}}Error{{end}}{{
 	if .IsScene }}  scene.BaseSceneError{ Scene: &{{ToTitle .Scene}}{TextManager: sc.TextManager} }{{end}}{{end}}`
 
-const directorConfigStruct = `
+const directorConfigStructT = `
 const GoodByeCommand = "{{ .GoodByeCommand }}"
 
 func New{{ ToTitle .Name }}Script(manager  manager.TextManager) game.SceneDirectorConfig {
 	return game.SceneDirectorConfig{
-		StartScene:   &{{ ToTitle .StartScene }}{manager},
-		GoodbyeScene: &{{ ToTitle .GoodByeScene }}{manager},
+		StartScene:   &scenes.{{ ToTitle .StartScene }}{
+						TextManager: manager,
+					},
+		GoodbyeScene: &scenes.{{ ToTitle .GoodByeScene }}{
+						TextManager: manager,
+					},
 		EndCommand:   GoodByeCommand,
 	}
 }
+`
+
+const scenesConstantConfigStructT = `{{ if .Script }}
+type SceneName string
+
+const ( {{ range $nameScene, $scene := .Script }}
+	// {{ ToTitle $nameScene }}Scene name of $nameScene
+	{{ ToTitle $nameScene }}Scene = SceneName("{{ ToSnake $nameScene }}") {{end}}
+	// NoScene name of next scene for ending scenes
+	NoScene = SceneName("_nothing_scene")
+) {{end}}
 `